+++ volmdlr/core_compiled.pyx
@@ -16,6 +16,7 @@ from typing import List, Text, Tuple
 import matplotlib.pyplot as plt
 import numpy as npy
 import plot_data
+import volmdlr
 from dessia_common.core import DessiaObject
 from matplotlib.patches import FancyArrow, FancyArrowPatch
 from mpl_toolkits.mplot3d import proj3d
@@ -1984,6 +1985,7 @@ class Point3D(Vector3D):
             and the new current id
         :rtype: tuple
         """
+        current_id += 1
         content = "#{} = CARTESIAN_POINT('{}',({:.6f},{:.6f},{:.6f}));\n".format(current_id, self.name,
                                                                                  1000. * self.x,
                                                                                  1000. * self.y,
@@ -3303,13 +3305,12 @@ class Frame3D(Basis3D):
         """
         content, origin_id = self.origin.to_point().to_step(current_id)
         current_id = origin_id + 1
+        w_content, w_id = Vector3D.to_step(self.w, current_id)
+        current_id = w_id + 1
         u_content, u_id = Vector3D.to_step(self.u, current_id)
         current_id = u_id + 1
-        v_content, v_id = Vector3D.to_step(self.v, current_id)
-        current_id = v_id + 1
-        content += u_content + v_content
-        content += "#{} = AXIS2_PLACEMENT_3D('{}',#{},#{},#{});\n"\
-            .format(current_id, self.name, origin_id, u_id, v_id)
+        content += w_content + u_content
+        content += f"#{current_id} = AXIS2_PLACEMENT_3D('{self.name}',#{origin_id},#{w_id},#{u_id});\n"
         return content, current_id
 
     def plot2d(self, x=X3D, y=Y3D, ax=None, color="k"):
@@ -3402,25 +3403,23 @@ class Frame3D(Basis3D):
         """
         origin = object_dict[arguments[1]]
         if arguments[2] == "$" and arguments[3] == "$":
-            u = Z3D
-            v = X3D
-        elif arguments[2] == "$":
-            frame = cls.from_point_and_vector(origin, object_dict[arguments[3]], main_axis=X3D)
-            u = frame.u
-            v = frame.v
-        elif arguments[3] == "$":
-            frame = cls.from_point_and_vector(origin, object_dict[arguments[2]], main_axis=X3D)
-            u = frame.u
-            v = frame.v
-        else:
-            u = object_dict[arguments[2]]
-            v = object_dict[arguments[3]]
-        w = u.cross(v)
+            return cls(origin, volmdlr.X3D, volmdlr.Y3D, volmdlr.Z3D, arguments[0][1:-1])
+        if arguments[2] == "$":
+            return cls.from_point_and_vector(origin, object_dict[arguments[3]], main_axis=X3D,
+                                             name= arguments[0][1:-1])
+        if arguments[3] == "$":
+            return cls.from_point_and_vector(origin, object_dict[arguments[2]], main_axis=Z3D,
+                                             name= arguments[0][1:-1])
+        w = object_dict[arguments[2]]
+        u = object_dict[arguments[3]]
+        u = u - u.dot(w) * w
+        u = u.unit_vector()
+        v = w.cross(u)
         return cls(origin, u, v, w, arguments[0][1:-1])
 
     @classmethod
     def from_point_and_vector(cls, point: Point3D, vector: Vector3D,
-                              main_axis: Vector3D = X3D):
+                              main_axis: Vector3D = X3D, name: str = ""):
         """
         Creates a new frame from a point and vector by rotating the global
         frame. Global frame rotates in order to have 'vector' and 'main_axis'
@@ -3436,6 +3435,8 @@ class Frame3D(Basis3D):
             (can be X3D, Y3D or Z3D). Default value is X3D,
             the vector (1, 0, 0)
         :type main_axis: :class:`volmdlr.Vector3D`, optional
+        :param name: Frame's name.
+        :type name: str
         :return: The created local frame
         :rtype: :class:`volmdlr.Frame3D`
         """
@@ -3465,7 +3466,7 @@ class Frame3D(Basis3D):
         v = Y3D.rotation(O3D, rot_axis, rot_angle)
         w = Z3D.rotation(O3D, rot_axis, rot_angle)
 
-        return cls(point, u, v, w)
+        return cls(point, u, v, w, name=name)
 
     @classmethod
     def from_3_points(cls, point1, point2, point3):