+++ volmdlr/surfaces.py
@@ -1089,20 +1089,20 @@ class Plane3D(Surface3D):
         :return: The corresponding Plane3D object.
         :rtype: :class:`volmdlr.faces.Plane3D`
         """
-        frame3d = object_dict[arguments[1]]
-        frame3d.normalize()
-        frame = volmdlr.Frame3D(frame3d.origin,
-                                      frame3d.v, frame3d.w, frame3d.u)
+        frame = object_dict[arguments[1]]
+        frame.normalize()
         return cls(frame, arguments[0][1:-1])
 
     def to_step(self, current_id):
         """
-        Transform a Plane 3D to step.
+        Converts the object to a STEP representation.
 
+        :param current_id: The ID of the last written primitive.
+        :type current_id: int
+        :return: The STEP representation of the object and the last ID.
+        :rtype: tuple[str, list[int]]
         """
-        frame = volmdlr.Frame3D(self.frame.origin, self.frame.w, self.frame.u,
-                                      self.frame.v)
-        content, frame_id = frame.to_step(current_id)
+        content, frame_id = self.frame.to_step(current_id)
         plane_id = frame_id + 1
         content += f"#{plane_id} = PLANE('{self.name}',#{frame_id});\n"
         return content, [plane_id]
@@ -2073,22 +2073,20 @@ class CylindricalSurface3D(PeriodicalSurface):
         """
 
         length_conversion_factor = kwargs.get("length_conversion_factor", 1)
-        frame3d = object_dict[arguments[1]]
-        u_vector, w_vector = frame3d.v, -frame3d.u
-        u_vector.normalize()
-        w_vector.normalize()
-        v_vector = w_vector.cross(u_vector)
-        frame_direct = volmdlr.Frame3D(frame3d.origin, u_vector, v_vector, w_vector)
+        frame = object_dict[arguments[1]]
         radius = float(arguments[2]) * length_conversion_factor
-        return cls(frame_direct, radius, arguments[0][1:-1])
+        return cls(frame, radius, arguments[0][1:-1])
 
     def to_step(self, current_id):
         """
-        Translate volmdlr primitive to step syntax.
+        Converts the object to a STEP representation.
+
+        :param current_id: The ID of the last written primitive.
+        :type current_id: int
+        :return: The STEP representation of the object and the last ID.
+        :rtype: tuple[str, list[int]]
         """
-        frame = volmdlr.Frame3D(self.frame.origin, self.frame.w, self.frame.u,
-                                      self.frame.v)
-        content, frame_id = frame.to_step(current_id)
+        content, frame_id = self.frame.to_step(current_id)
         current_id = frame_id + 1
         content += f"#{current_id} = CYLINDRICAL_SURFACE('{self.name}',#{frame_id},{round(1000 * self.radius, 4)});\n"
         return content, [current_id]
@@ -2410,36 +2408,21 @@ class ToroidalSurface3D(PeriodicalSurface):
 
         length_conversion_factor = kwargs.get("length_conversion_factor", 1)
 
-        frame3d = object_dict[arguments[1]]
-        u_vector, w_vector = frame3d.v, -frame3d.u
-        u_vector.normalize()
-        w_vector.normalize()
-        v_vector = w_vector.cross(u_vector)
-        frame_direct = volmdlr.Frame3D(frame3d.origin, u_vector, v_vector, w_vector)
+        frame = object_dict[arguments[1]]
         rcenter = float(arguments[2]) * length_conversion_factor
         rcircle = float(arguments[3]) * length_conversion_factor
-        return cls(frame_direct, rcenter, rcircle, arguments[0][1:-1])
+        return cls(frame , rcenter, rcircle, arguments[0][1:-1])
 
     def to_step(self, current_id):
         """
         Converts the object to a STEP representation.
 
-        This method converts the object to a STEP (Standard for the Exchange of Product model data) representation.
-        It first creates a copy of the object's frame using the `volmdlr.Frame3D` constructor. Then, it converts
-        the frame to a STEP representation using the `to_step` method of the `Frame3D` class. The resulting content
-        and updated ID are stored in the `content` variable and the `current_id` variable, respectively. Next,
-        the method constructs a STEP string representing the toroidal surface using the object's attributes and
-        the generated frame ID. The content is updated with the toroidal surface representation. Finally, the
-        updated content and a list containing the updated current ID are returned as a tuple.
-
-        :param current_id: The current ID counter for generating unique STEP entity IDs.
-
-        :return: A tuple containing the STEP representation of the object and the updated current ID.
+        :param current_id: The ID of the last written primitive.
+        :type current_id: int
+        :return: The STEP representation of the object and the last ID.
+        :rtype: tuple[str, list[int]]
         """
-
-        frame = volmdlr.Frame3D(self.frame.origin, self.frame.w, self.frame.u,
-                                      self.frame.v)
-        content, frame_id = frame.to_step(current_id)
+        content, frame_id = self.frame.to_step(current_id)
         current_id = frame_id + 1
         content += f"#{current_id} = TOROIDAL_SURFACE('{self.name}',#{frame_id}," \
                    f"{round(1000 * self.tore_radius, 4)},{round(1000 * self.small_radius, 4)});\n"
@@ -2745,6 +2728,9 @@ class ConicalSurface3D(PeriodicalSurface):
         PeriodicalSurface.__init__(self, frame=frame, name=name)
 
     def plot(self, ax=None, color='grey', alpha=0.5, **kwargs):
+        """
+        Plots the ConicalSurface3D.
+        """
         z = kwargs.get("z", 0.5)
         if ax is None:
             fig = plt.figure()
@@ -2779,21 +2765,22 @@ class ConicalSurface3D(PeriodicalSurface):
         length_conversion_factor = kwargs.get("length_conversion_factor", 1)
         angle_conversion_factor = kwargs.get("angle_conversion_factor", 1)
 
-        frame3d = object_dict[arguments[1]]
-        u, w = frame3d.v, frame3d.u
-        u.normalize()
-        w.normalize()
-        v = w.cross(u)
+        frame = object_dict[arguments[1]]
         radius = float(arguments[2]) * length_conversion_factor
         semi_angle = float(arguments[3]) * angle_conversion_factor
-        origin = frame3d.origin - radius / math.tan(semi_angle) * w
-        frame_direct = volmdlr.Frame3D(origin, u, v, w)
-        return cls(frame_direct, semi_angle, arguments[0][1:-1])
+        frame.origin = frame.origin - radius / math.tan(semi_angle) * frame.w
+        return cls(frame, semi_angle, arguments[0][1:-1])
 
     def to_step(self, current_id):
-        frame = volmdlr.Frame3D(self.frame.origin, self.frame.w, self.frame.u,
-                                      self.frame.v)
-        content, frame_id = frame.to_step(current_id)
+        """
+        Converts the object to a STEP representation.
+
+        :param current_id: The ID of the last written primitive.
+        :type current_id: int
+        :return: The STEP representation of the object and the last ID.
+        :rtype: tuple[str, list[int]]
+        """
+        content, frame_id = self.frame.to_step(current_id)
         current_id = frame_id + 1
         content += f"#{current_id} = CONICAL_SURFACE('{self.name}',#{frame_id},{0.},{round(self.semi_angle, 4)});\n"
         return content, [current_id]
@@ -2880,8 +2867,12 @@ class ConicalSurface3D(PeriodicalSurface):
             return []
         start3d = self.point2d_to_3d(linesegment2d.start)
         center = self.frame.origin + param_z1 * self.frame.w
-        circle = curves.Circle3D(volmdlr.Frame3D(
-            center, self.frame.u, self.frame.v, self.frame.w), center.point_distance(start3d))
+        if linesegment2d.unit_direction_vector().dot(volmdlr.X2D) > 0:
+            circle = curves.Circle3D(volmdlr.Frame3D(
+                center, self.frame.u, self.frame.v, self.frame.w), center.point_distance(start3d))
+        else:
+            circle = curves.Circle3D(volmdlr.Frame3D(
+                center, self.frame.u, -self.frame.v, -self.frame.w), center.point_distance(start3d))
         if math.isclose(param_z1, param_z2, abs_tol=1e-4):
             if abs(theta1 - theta2) == volmdlr.TWO_PI:
                 return [edges.FullArc3D(circle, start3d)]
@@ -3078,22 +3069,20 @@ class SphericalSurface3D(PeriodicalSurface):
         """
         length_conversion_factor = kwargs.get("length_conversion_factor", 1)
 
-        frame3d = object_dict[arguments[1]]
-        u_vector, w_vector = frame3d.v, frame3d.u
-        u_vector.normalize()
-        w_vector.normalize()
-        v_vector = w_vector.cross(u_vector)
-        frame_direct = volmdlr.Frame3D(frame3d.origin, u_vector, v_vector, w_vector)
+        frame = object_dict[arguments[1]]
         radius = float(arguments[2]) * length_conversion_factor
-        return cls(frame_direct, radius, arguments[0][1:-1])
+        return cls(frame, radius, arguments[0][1:-1])
 
     def to_step(self, current_id):
         """
-        Translate volmdlr primitive to step syntax.
+        Converts the object to a STEP representation.
+
+        :param current_id: The ID of the last written primitive.
+        :type current_id: int
+        :return: The STEP representation of the object and the last ID.
+        :rtype: tuple[str, list[int]]
         """
-        frame = volmdlr.Frame3D(self.frame.origin, self.frame.w, self.frame.u,
-                                      self.frame.v)
-        content, frame_id = frame.to_step(current_id)
+        content, frame_id = self.frame.to_step(current_id)
         current_id = frame_id + 1
         content += f"#{current_id} = SPHERICAL_SURFACE('{self.name}',#{frame_id},{round(1000 * self.radius, 4)});\n"
         return content, [current_id]
@@ -3151,6 +3140,8 @@ class SphericalSurface3D(PeriodicalSurface):
         start = self.point2d_to_3d(linesegment2d.start)
         interior = self.point2d_to_3d(0.5 * (linesegment2d.start + linesegment2d.end))
         end = self.point2d_to_3d(linesegment2d.end)
+        if start.is_close(interior) and interior.is_close(end) and end.is_close(start):
+            return []
         u_vector = start - self.frame.origin
         u_vector.normalize()
         v_vector = interior - self.frame.origin